Extract `ChatMessage` to its own file.
